Remove abstract CoordinateReader
